requirements.txt
# Web framework
flask==3.1.0

# Google Gemini API
google-generativeai==0.8.5

# Security: restricted Python execution
RestrictedPython==7.4

# Environment variable management
python-dotenv==1.0.1
